Database: Remove AnotherBlueTheme
This is a clear violation of Google Play's Malicious Behavior: https://play.google.com/about/privacy-security/malicious-behavior/index.html

"Apps or SDKs that download executable code, such as dex files or native code, from a source other than Google Play."

Since this is a direct download. It could be considered a gray area because Chrome would actually be doing the downloading but we are still providing a direct link. Better be safe than sorry given that Google is cracking down on security lately, as seen with the Magisk situation. Ice is also another one that may be in violation of this given we are linking to it but we can deal with that one later.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>